Задание 12. Работа со строкой запроса (после упрощения иерархии)
