+++ lot.php
@@ -1,44 +1,38 @@
 <?php
 
-    require_once('../constants.php');
-    require_once('../connection.php');
-    require_once('../functions.php');
-    require_once('../query_functions.php');
+    require_once('constants.php');
+    require_once('connection.php');
+    require_once('functions.php');
+    require_once('db_functions.php');
 
     $lot_id = isset($_GET['id']) ? $_GET['id'] : null;
-    $lots = $lot_id ? get_lot_info($connection, $lot_id) : [];
-    $lot = isset($lots[0]) && !was_error($lots) ? $lots[0] : null;
+    $lot = $lot_id ? get_lot_info($connection, $lot_id) : null;
+
+    $is_ok = ($lot_id && $lot && !was_error($lot));
 
     $categories = get_all_categories($connection);
 
     $categories_content = include_template('categories.php',
         [
             'categories' => $categories,
-            'ul_classname' => 'nav__list container',
-            'li_classname' => 'nav__item',
-            'a_classname' => ''
+            'style' => get_assoc_element($category_styles, 'bar')
         ],
         true);
 
-    if ($lot_id && $lot) {
-
-        $page_content = include_template('lot.php',
+    $page_content = $is_ok ?
+        include_template('lot.php',
             [
                 'lot' => $lot,
                 'categories_content' => $categories_content,
-                'images' => '../' . get_assoc_element($paths, 'images'),
+                'images' => get_assoc_element($paths, 'images'),
             ],
-            true);
-
-    } else {
-
-        $page_content = include_template('404.php',
+            true)
+        :
+        include_template('404.php',
             [
                 'categories_content' => $categories_content,
             ],
             true);
-
-    }
 
     $layout_content = include_template('layout.php',
         [
@@ -49,5 +43,9 @@
             'user_name' => $user_name
         ],
         true);
+
+    if (!$is_ok) {
+        http_response_code(404);
+    }
 
     print($layout_content);
