+++ templates/index.php
@@ -22,7 +22,7 @@
                         <div class="lot__info">
                             <span class="lot__category"><?= strip_tags($lot['category']); ?></span>
                             <h3 class="lot__title">
-                                <a class="text-link" href="pages/lot.php?id=<?= strip_tags($lot['id']); ?>">
+                                <a class="text-link" href="lot.php?id=<?= strip_tags($lot['id']); ?>">
                                     <?= strip_tags($lot['name']); ?>
                                 </a>
                             </h3>
