+++ templates/layout.php
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title><?= $title; ?></title>
-    <link href="../css/normalize.min.css" rel="stylesheet">
-    <link href="../css/style.css" rel="stylesheet">
+    <link href="css/normalize.min.css" rel="stylesheet">
+    <link href="css/style.css" rel="stylesheet">
 </head>
 <body>
 <div class="page-wrapper">
